+++ packages/dapp-kit-react-privy/src/components/AccountModal/Components/index.ts
@@ -1,2 +1,3 @@
 export * from './AccountDetailsButton';
 export * from './ActionButton';
+export * from './AccountSelector';
