+++ packages/dapp-kit-react-privy/src/components/ConnectModal/Contents/MainContent.tsx
@@ -19,18 +19,18 @@ import {
     useLoginWithOAuth,
     useLoginWithPasskey,
 } from '@privy-io/react-auth';
-import { useDAppKitPrivyConfig } from '../../providers/DAppKitPrivyProvider';
-import { FadeInViewFromBottom, VersionFooter } from '../common';
+import { useDAppKitPrivyConfig } from '../../../providers/DAppKitPrivyProvider';
+import { FadeInViewFromBottom, VersionFooter } from '../../common';
 import { HiOutlineWallet } from 'react-icons/hi2';
 import { FcGoogle } from 'react-icons/fc';
-import { VechainLogo } from '../../assets';
+import { VechainLogo } from '../../../assets';
 import { CiCircleMore } from 'react-icons/ci';
-import { ConnectModalContents } from './ConnectModal';
+import { ConnectModalContents } from '../ConnectModal';
 import { IoIosApps, IoIosFingerPrint } from 'react-icons/io';
 import { useWalletModal } from '@vechain/dapp-kit-react';
-import { VECHAIN_PRIVY_APP_ID } from '../../utils';
+import { VECHAIN_PRIVY_APP_ID } from '../../../utils';
 import { useEffect } from 'react';
-import { useWallet } from '../../hooks';
+import { useWallet } from '../../../hooks';
 
 type Props = {
     setCurrentContent: React.Dispatch<
