+++ packages/dapp-kit-react-privy/src/components/common/index.ts
@@ -1,10 +1,5 @@
-export * from './FadeInViewFromRight';
-export * from './FadeInViewFromLeft';
-export * from './FadeInViewFromBottom';
+export * from './Motion';
 export * from './ModalBackButton';
 export * from './AddressDisplay';
-export * from './AppLogos';
-export * from './AccountSelector';
-export * from './FadeInView';
 export * from './VersionFooter';
 export * from './StickyHeaderContainer';
