+++ packages/dapp-kit-react-privy/src/utils/index.ts
@@ -1,5 +1,5 @@
 export * from './Constants';
 export * from './FormattingUtils';
-export * from './SmartAccountTransactionForwarder';
+export * from './RandomTxForwarder';
 export * from './Types';
 export * from './Constants';
